refactor(@ngtools/webpack): initial ivy-only Webpack plugin/loader
This change introduces a new Ivy-only Webpack plugin.  The plugin is based on the design introduced by `tsc_wrapped` within Bazel’s `rules_typescript` and leverages the newly added ngtsc plugin from within the `@angular/compiler-cli` package.  By leveraging the same plugin interface that it used by Bazel, a common interface can be used to access the Angular compiler.  This has benefits to both major build systems and dramatically reduces the necessary code infrastructure to integrate the Angular compiler.
The plugin also simplifies and reduces the amount of code within the plugin by leveraging newer TypeScript features and capabilities.  The need for the virtual filesystem has also been removed.  The file replacements capability was the primary driver for the previous need for the virtual filesystem.  File replacements are now implemented using a two-pronged approach.  The first, for TypeScript, is to hook TypeScript module resolution and adjust the resolved modules based on the configured file replacements.  This is similar in behavior to TypeScript path mapping.  The second, for Webpack, is the use of the `NormalModuleReplacementPlugin` to facilitate bundling of the configured file replacements.  An advantage to this approach is that the build system (both TypeScript and Webpack) are now aware of the replacements and can operate without augmenting multiple aspects of system as was needed previously.
The plugin also introduces the use of TypeScript’s builder programs.  The current primary benefit is more accurate and simplified dependency discovery.  Further, they also provide for the potential future introduction of incremental build support and incremental type checking.

NOTE: The deprecated string format for lazy routes is not supported by this plugin.  Dynamic imports are recommended for use with Ivy and are required when using the new plugin.