feat(@angular-devkit/build-angular): support browserslist environments via configurations
The project configuration name can now be used as a browserslist environment when configuring the `.browserslistrc` file.  This allows for differing the list of supported browsers based on the configuration used to build the application.  For example, browserslist willl use the `production` environment, if present, when the `production` build configuration is used.
Please note that the default "development" build configuration is unnamed and Browserslist defaults to using the `production` environment if not specified.  As a result, the `production` environment will be used when not specifying a build configuration.
Browserslist environment documentation: https://github.com/browserslist/browserslist#configuring-for-different-environments